Добавлена обработка исключений при поиске id текущего пользователся

+++ AspCourse/Models/database/entity/Location.cs
@@ -7,13 +7,13 @@ namespace AspCourse.Models
 {
     public class Location
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
 
         public Tournament Tournament { get; set; }
 
-        public long? UserId { get; set; }
+        public int? UserId { get; set; }
         public User User { get; set; }
 
         public Location()
